Move epoch serializer to the custom serde module
- Fix issue deserializing i64
- Add test for the specific date
